Split tests by adapters.
Run CI/CD tests per adapter with own Gemfiles.
Allow to test in future single adpater per changed files.
Moved adapters tests to seprate folder.
Update dependabot to check custom gemfiles.
Add github action to update gems.
