+++ .yamllint.yml
@@ -17,7 +17,7 @@ rules:
   indentation:
     spaces: 2
     indent-sequences: true
-    check-multi-line-strings: true
+    check-multi-line-strings: false
   line-length:
     max: 100
     level: warning
