+++ gemfiles/net_http.gemfile
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+source "https://rubygems.org"
+
+gem "rake"
+gem "rake-compiler"
+
+gem "hiredis-client", github: "redis-rb/redis-client"
+gem "hiredis", "~> 0.6"
+gem "minitest"
+gem "mocha"
+gem "redis", "~> 4.8"
+gem "timecop"
+gem "toxiproxy"
+gem "webrick"
+
+gemspec path: "../"
