+++ gemfiles/redis_4.gemfile
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+source "https://rubygems.org"
+
+gem "rake"
+gem "rake-compiler"
+gem "minitest"
+gem "mocha"
+gem "timecop"
+gem "toxiproxy"
+gem "webrick"
+
+gem "hiredis-client"
+gem "hiredis", "~> 0.6"
+gem "redis", "~> 4.8"
+
+gemspec path: "../"
