+++ gemfiles/redis_5.gemfile.lock
@@ -1,0 +1,44 @@
+PATH
+  remote: ..
+  specs:
+    semian (0.13.3)
+
+GEM
+  remote: https://rubygems.org/
+  specs:
+    connection_pool (2.2.5)
+    hiredis (0.6.3)
+    hiredis-client (0.7.4)
+      redis-client (= 0.7.4)
+    minitest (5.16.3)
+    mocha (1.14.0)
+    rake (13.0.6)
+    rake-compiler (1.2.0)
+      rake
+    redis (5.0.3)
+      redis-client (>= 0.7.4)
+    redis-client (0.7.4)
+      connection_pool
+    timecop (0.9.5)
+    toxiproxy (2.0.2)
+    webrick (1.7.0)
+
+PLATFORMS
+  aarch64-linux
+  arm64-darwin-21
+
+DEPENDENCIES
+  hiredis (~> 0.6)
+  hiredis-client
+  minitest
+  mocha
+  rake
+  rake-compiler
+  redis
+  semian!
+  timecop
+  toxiproxy
+  webrick
+
+BUNDLED WITH
+   2.3.14
