+++ gemfiles/redis_client.gemfile
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+source "https://rubygems.org"
+
+gem "rake"
+gem "rake-compiler"
+gem "minitest"
+gem "mocha"
+gem "timecop"
+gem "toxiproxy"
+gem "webrick"
+
+gem "hiredis-client", github: "redis-rb/redis-client"
+gem "hiredis", "~> 0.6"
+gem "redis", github: "redis/redis-rb"
+
+gemspec path: "../"
