+++ gemfiles/redis_client.gemfile.lock
@@ -1,0 +1,54 @@
+GIT
+  remote: https://github.com/redis-rb/redis-client.git
+  revision: a03107a80516f053b1909cde04a6d8f6ef5805e8
+  specs:
+    hiredis-client (0.7.4)
+      redis-client (= 0.7.4)
+    redis-client (0.7.4)
+      connection_pool
+
+GIT
+  remote: https://github.com/redis/redis-rb.git
+  revision: 2e9ab3834caf661a449b4b8376deb65713ee4fcf
+  specs:
+    redis (5.0.3)
+      redis-client (>= 0.7.4)
+
+PATH
+  remote: ..
+  specs:
+    semian (0.13.3)
+
+GEM
+  remote: https://rubygems.org/
+  specs:
+    connection_pool (2.2.5)
+    hiredis (0.6.3)
+    minitest (5.16.3)
+    mocha (1.14.0)
+    rake (13.0.6)
+    rake-compiler (1.2.0)
+      rake
+    timecop (0.9.5)
+    toxiproxy (2.0.2)
+    webrick (1.7.0)
+
+PLATFORMS
+  aarch64-linux
+  arm64-darwin-21
+
+DEPENDENCIES
+  hiredis (~> 0.6)
+  hiredis-client!
+  minitest
+  mocha
+  rake
+  rake-compiler
+  redis!
+  semian!
+  timecop
+  toxiproxy
+  webrick
+
+BUNDLED WITH
+   2.3.21
