+++ test/adapters/grpc/echo_pb.rb
