+++ test/adapters/grpc/echo_service.rb
