+++ test/adapters/net_http_test.rb
