+++ test/test_helper.rb
@@ -4,17 +4,11 @@ require "rubygems"
 require "bundler/setup"
 
 require "minitest/autorun"
-require "active_record"
-require "mysql2"
 require "semian"
-require "semian/mysql2"
-require "semian/redis"
-require "semian/redis_client"
 require "toxiproxy"
 require "timecop"
 require "tempfile"
 require "fileutils"
-require "byebug"
 require "mocha"
 require "mocha/minitest"
